requirements.txt
requests
slixmpp
google
google.genai
markdown-it-py
